rtl8xxxu: force resetting basic rate to all except CCK.
There might be something wrong with the set_basic_rate callback
which cause the driver only ack the packet with specific tx rate
(MCS2 19.5Mbps). It results in the dramatically slow RX throughtput.

In this dirty hack, we force the response rate with all rates except
CCK 2, 5.5, and 11Mbps as vendor driver does.