rtl8xxxu: move the arg to be the last element of the h2c command
From the code of the rtl8192cu and vendor driver, the rate mask
set command is composed by a 5-elements array and the arg0 is
the last element. Move the arg0 to the last until we vefiry the
rtl8xxxu_gen1_h2c_cmd is doing the same as _rtl92c_fill_h2c_command.