rtl8xxxu: Improve the retransmission rate with HW_RTS enable
From the RTL8188CUS vendor driver, the HW_RTS is enabled when
ampdu enabled. The retransmission rate is about ~20% even in a
very clean environment with rtl8xxxu driver compared to ~8%
with the vendor driver.

This commit tries to enable the HW_RTS and succeeds to improve
the retransmission rate from ~20% to ~12%. Need to apply the same
thing in rtl8192cu driver to make sure whether it really improves
or not.

Signed-Off-By: Chris Chiu <chiu@endlessos.org>