rtl8xxxu: Improve the retransmission rate on RX AMPDU
From the RTL8188CUS vendor driver and upstream rtw88 driver, the
init value for Response Rate Set Register is 0x15f/0x150. This will
affect the tx rate of the block ack packet. If we leave the register
to be the default value 0xf, the block ack will always be transmitted
with n-rate instead of g-rate and it will never trigger the AP to
send RTS-CTS to protect the A-MPDU aggregation. The retransmission
rate will the get higher than expected.

Signed-Off-By: Chris Chiu <chrischiu@canonical.com>