ASoC: cs35l56: Add control to read CAL_SET_STATUS
BugLink: https://bugs.launchpad.net/bugs/2139391

Create an ALSA control to read the value of the firmware
CAL_SET_STATUS control. This reports whether the firmware is
using a calibration blob or the default calibration from the
.bin file.

The firmware only reports a valid value in this register while
audio is actually playing and the internal PLL is locked to the
audio clock. Otherwise it returns a status of "unknown".

Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>
Link: https://patch.msgid.link/20251111130850.513969-2-rf@opensource.cirrus.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 69f3474a01e9867dd99fc4b703fa834ea1835c7d)
Signed-off-by: Chris Chiu <chris.chiu@canonical.com>